+++ pageContent/BlogSlugPage.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variable';
+@import '../styles/variable';
 
 .wrap {
   display: flex;
